Implemented semi-working ping in JavaScript.
This doesn't really work as intended. If we're connected to the Internet, we can measure our latency when pinging www.google.com, and we'll time out after 10 seconds. If we're disconnected (ie. go into airplane mode on phone with Bluetooth on), our ping functionality resolves to 'error' right away instead of timing out, which reports a very low latency to the Pebble. I may need to use an alternate ping functionality (eg. using XMLHttpRequest) instead of the Image object hack currently being used.