JavaScript ping works. Pebble C strings updated.
We still request a fake image from Google and measure the latency, then send that to the Pebble. We use an XMLHttpRequest; instead of using an 'img.onerror' property, we know we're connected if we get a 404/other valid error code, and we're disconnected or timed out if we get error code 0. Currently tested on regular webserver (normal), webserver that takes a long time to respond (timeout), and in airplane mode (disconnected).