Drop keypresses during ping, Bluetooth timeout.
If Bluetooth on our phone or watch is turned off, we don't want our app to hang and say "Pinging..." for an infinitely long time; time out if we don't get a response within 1.5x the ping timeout period.

Additionally, if we can 'spam' requests while a ping is still going on, we get increased complexity in our .js code, and wasted battery via Bluetooth for little to no benefit. Simply drop keypresses while a ping is ongoing / before the Bluetooth timeout period ends.