Added wifi animation that shows during ping.
We show an animation so that the user knows we're not simply hanging / in an infinite loop, and are still waiting for a response from the phone. This isn't actually an Animation from the Pebble API; we simply update the .png in the image layer periodically. This appears to be the simplest way to implement this, but there should be methods of only updating the part of the screen we're interested in. This should be more power-efficient. This is an optimization I may make later.